refactor: :truck: move theme provider to context directory
